+++ test/unit/vagrant/action/builtin/box_add_test.rb
@@ -164,6 +164,19 @@ describe Vagrant::Action::Builtin::BoxAdd do
         to raise_error(Vagrant::Errors::BoxChecksumMismatch)
     end
 
+    it "raises an error if the box path doesn't exist" do
+      box_path = iso_env.box2_file(:virtualbox)
+
+      env[:box_name] = "foo"
+      env[:box_url] = box_path.to_s + "nope"
+
+      box_collection.should_receive(:add).never
+      app.should_receive(:call).never
+
+      expect { subject.call(env) }.
+        to raise_error(Vagrant::Errors::DownloaderError)
+    end
+
     it "force adds if exists and specified" do
       box_path = iso_env.box2_file(:virtualbox)